requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-celery-results==2.1.0
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
idna==2.10
importlib-metadata==4.5.0
kombu==5.1.0
prompt-toolkit==3.0.18
PyJWT==1.7.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
